requirements.txt
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
